Add files "StaffMember.cs" and "Person.cs". Add StaffMember menu,Add function "AddStaffMember"
Person is abstract class which is inherited by classed Person and StaffMember.
Beacuse of AddStaffMember function "YearInput" function is fixed to check if new person is too old or too young(under 18).
Functions "ObjectExists" and "IsDefinedInENum" are migrated to Helper class so they can be made generic to fit any type of private list or enum.